Follow DRY.
